Create Code: Insert To button With find
This was added to the slasher project as an enhancement. 
The insert button will insert " to " into text replacing the selected vergule.
The lower With function searches the document for hte next vergule, an enhancement which will help prevent the random insertion of text in the document. 
The With function prevents the user step of clicking the "Find Next /" button.